pyproject.toml
[project]
name = "agent-gateway"
version = "0.1.0"
description = "Modular OpenAI-compatible gateway with agent registry, tools, and LM Studio upstreams"
authors = [{ name = "Agent Gateway Team" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.5,<0.116",
  "uvicorn[standard]>=0.30.6,<0.31",
  "openai>=1.52.2,<2.0",
  "pyyaml>=6.0.2,<7.0",
  "pydantic>=2.9.2,<3.0",
  "prometheus-client>=0.20.0,<0.21",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "httpx>=0.27.2,<0.28",
  "pytest-cov>=5.0,<6.0",
  "ruff>=0.6.5,<0.7",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = [
  "api*",
  "agents*",
  "registry*",
  "tooling*",
  "sdk_adapter*",
  "config*",
  "observability*",
  "security*",
]


